+++ src/ideabox/policy/faceted/subscriber.py
@@ -1,8 +1,11 @@
 # -*- coding: utf-8 -*-
 
-from datetime import datetime
 from eea.facetednavigation.interfaces import IFacetedLayout
 from hashlib import md5
+from ideabox.policy import utils
+from plone.memoize import ram
+
+import time
 
 
 def faceted_query_handler(obj, event):
@@ -12,28 +15,51 @@ def faceted_query_handler(obj, event):
         except TypeError:
             layout = None
         if layout == "faceted-project":
-            event.query["sort_on"] = Randomizer(obj.REQUEST).random_sort_key
-            event.query["sort_order"] = "ascending"
+            randomizer = Randomizer(obj.REQUEST)
+            event.query["sort_on"] = randomizer.random_sort_key
+            event.query["sort_order"] = randomizer.random_sort_key
+
+
+def _cache_key(function, generator):
+    return (generator, time.time() // (12 * 60 * 60))
+
+
+@ram.cache(_cache_key)
+def random_key_generator(generator):
+    return md5(generator).hexdigest()
 
 
 class Randomizer(object):
 
-    _keys = {0: "UID", 1: "sortable_title", 2: "created", 3: "modified"}
+    _keys = {
+        0: "UID",
+        1: "project_random_1",
+        2: "project_random_2",
+        3: "project_random_3",
+        4: "project_random_4",
+        5: "project_random_5",
+        6: "project_random_6",
+        7: "project_random_7",
+    }
+    _sort_order_keys = ("ascending", "descending")
 
     def __init__(self, request):
         self.request = request
+        now = utils.now()
+        generator = "{0}-{1}".format(
+            self.request.cookies.get(
+                "__ac", self.request.cookies.get("beaker.session")
+            ),
+            now.strftime("%H"),
+        )
+        self.key = random_key_generator(generator)
 
     @property
     def random_sort_key(self):
-        key = ord(self.random_key[:1])
+        key = ord(self.key[0])
         return self._keys[key % len(self._keys)]
 
     @property
-    def random_key(self):
-        now = datetime.now()
-        return md5(
-            u"{0}-{1}".format(
-                self.request.get("HTTP_USER_AGENT", now.strftime("%m%Y")),
-                now.strftime("%d"),
-            )
-        ).hexdigest()
+    def random_sort_order(self):
+        key = ord(self.key[-1])
+        return self._sort_order_keys[key % len(self._sort_order_keys)]